Moved Earning to Server from Mock
Moved Earning to Server from Mock, updated references
