+++ src/mock/Report.java
@@ -3,6 +3,7 @@ package mock;
 import java.util.ArrayList;
 import java.util.Date;
 import interfaces.ReportInterface;
+import server.Earning;
 
 public class Report implements ReportInterface {
 	private static int count = 0;
